[CDSR-1333][AO] remove code duplication from computing common rejected goods journey properties
